+++ bin/gtfsrealtime-update.js
@@ -2,10 +2,13 @@
 
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
+import PrettyError from 'pretty-error';
 
 import { getConfig } from '../lib/file-utils.js';
 import { formatError } from '../lib/log-utils.js';
 import { updateGtfsRealtime } from '../lib/gtfs.js';
+
+const pe = new PrettyError();
 
 const { argv } = yargs(hideBin(process.argv))
   .usage('Usage: $0 --configPath ./config.json')
@@ -19,7 +22,7 @@ const { argv } = yargs(hideBin(process.argv))
 const handleError = (error) => {
   const text = error || 'Unknown Error';
   process.stdout.write(`\n${formatError(text)}\n`);
-  console.error(error);
+  console.error(pe.render(error));
   process.exit(1);
 };